# [Numpy] OP_interp