… src/operator/numpy/np_interp_op.cu
… tests/python/unittest/test_numpy_interoperability.py
… tests/python/unittest/test_numpy_op.py