@Tommliu 2020-03-25
@mxnet-bot run ci [ci/jenkins/mxnet-validation/centos-gpu, ci/jenkins/mxnet-validation/unix-cpu,  ci/jenkins/mxnet-validation/windows-gpu]