@Tommliu 2020-03-25
@mxnet-bot run ci [centos-gpu, unix-cpu, windows-gpu]